Revert "Merge branch 'mercedesme-vehilce-sync' of https://github.com/weymann/openhab-addons into mercedesme-vehilce-sync"
This reverts commit 31677357423e78eb151655f691e8a2ae19e8ea0a, reversing
changes made to 081454fae1404f1f304934e0b008f74a75d594f5.

+++ bundles/org.openhab.binding.knx/src/main/java/org/openhab/binding/knx/internal/config/BridgeConfiguration.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2010-2023 Contributors to the openHAB project
+ * Copyright (c) 2010-2022 Contributors to the openHAB project
  *
  * See the NOTICE file(s) distributed with this work for additional
  * information.
@@ -12,34 +12,36 @@
  */
 package org.openhab.binding.knx.internal.config;
 
+import java.math.BigDecimal;
+
 import org.eclipse.jdt.annotation.NonNullByDefault;
 
 /**
  * {@link org.openhab.binding.knx.internal.handler.KNXBridgeBaseThingHandler} configuration
  *
- * @author Simon Kaufmann - Initial contribution
+ * @author Simon Kaufmann - initial contribution and API
  *
  */
 @NonNullByDefault
 public class BridgeConfiguration {
     private int autoReconnectPeriod = 0;
-    private int readingPause = 0;
-    private int readRetriesLimit = 0;
-    private int responseTimeout = 0;
+    private BigDecimal readingPause = BigDecimal.valueOf(0);
+    private BigDecimal readRetriesLimit = BigDecimal.valueOf(0);
+    private BigDecimal responseTimeout = BigDecimal.valueOf(0);
 
     public int getAutoReconnectPeriod() {
         return autoReconnectPeriod;
     }
 
-    public int getReadingPause() {
+    public BigDecimal getReadingPause() {
         return readingPause;
     }
 
-    public int getReadRetriesLimit() {
+    public BigDecimal getReadRetriesLimit() {
         return readRetriesLimit;
     }
 
-    public int getResponseTimeout() {
+    public BigDecimal getResponseTimeout() {
         return responseTimeout;
     }
