+++ bundles/org.openhab.binding.knx/src/main/java/org/openhab/binding/knx/internal/config/DeviceConfig.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2010-2023 Contributors to the openHAB project
+ * Copyright (c) 2010-2022 Contributors to the openHAB project
  *
  * See the NOTICE file(s) distributed with this work for additional
  * information.
@@ -12,20 +12,23 @@
  */
 package org.openhab.binding.knx.internal.config;
 
+import java.math.BigDecimal;
+
 import org.eclipse.jdt.annotation.NonNullByDefault;
 
 /**
  * Configuration object for the device thing handler.
  *
  * @author Karel Goderis - Initial contribution
- * @author Simon Kaufmann - refactoring and cleanup
+ * @author Simon Kaufmann - refactoring & cleanup
  */
 @NonNullByDefault
 public class DeviceConfig {
+
     private String address = "";
     private boolean fetch = false;
-    private int pingInterval = 0;
-    private int readInterval = 0;
+    private BigDecimal pingInterval = BigDecimal.valueOf(0);
+    private BigDecimal readInterval = BigDecimal.valueOf(0);
 
     public String getAddress() {
         return address;
@@ -35,11 +38,11 @@ public class DeviceConfig {
         return fetch;
     }
 
-    public int getPingInterval() {
+    public BigDecimal getPingInterval() {
         return pingInterval;
     }
 
-    public int getReadInterval() {
+    public BigDecimal getReadInterval() {
         return readInterval;
     }
 }
